inventory table decimal changes

+++ SriLankanParadise.ERP.UserManagement/ERP Web/Models/RequestModels/LocationInventoryRequestModel.cs
@@ -9,7 +9,7 @@
         public int LocationId { get; set; }
         public int MovementTypeId { get; set; }
 
-        public int? StockInHand { get; set; }
+        public decimal? StockInHand { get; set; }
         public string? BatchNo { get; set; }
 
         public int PermissionId { get; set; }
